+++ mod/twitch/src/main/java/tv/twitch/android/shared/bits/BitsSpendingPresenter.java
@@ -1,0 +1,7 @@
+package tv.twitch.android.shared.bits;
+
+import androidx.fragment.app.FragmentActivity;
+
+public class BitsSpendingPresenter {
+    public FragmentActivity activity;
+}